Reorganise tooltip content code
